# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для учета на предприятии /ООО Ромашка/
  version: "1.0"
  title: RestApi

produces:
  - application/json
  - application/xml
  - application/text

basePath: "/api"

definitions:
  MeasurementUnitModel:
    type: object
    description: "Единица измерения"
    properties:
      uid:
        type: string
        description: "Идентификатор единицы измерения"
      name:
        type: string
        description: "Название единицы измерения"
      base_unit:
        type: object
        description: "Базовая единица измерения"
        $ref: "#/definitions/MeasurementUnitModel"

  NomenclatureModel:
    type: object
    properties:
      uid:
        type: string
        description: "Идентификатор номенклатуры"
      name:
        type: string
        description: "Название номенклатуры"
      nomenclature_group:
        type: object
        description: "Группа номенклатуры"
        properties:
          uid:
            type: string
            description: "Идентификатор группы номенклатуры"
          name:
            type: string
            description: "Название группы номенклатуры"
      measurement_unit:
        type: object
        description: "Единица измерения"
        properties:
          uid:
            type: string
            description: "Идентификатор единицы измерения"
          name:
            type: string
            description: "Название единицы измерения"
          base_unit:
            type: object
            description: "Базовая единица измерения"
            $ref: "#/definitions/MeasurementUnitModel"

  Filter:
    type: object
    properties:
      field:
        type: string
        description: "Имя фильтруемого поля"
      value:
        type: string
        description: "Искомое значение"
      type:
        type: string
        enum:
          - EQUAL
          - LIKE

  PeriodFilter:
    type: object
    properties:
      date:
        type: string
        description: "Параметр даты"
      filter_type:
        description: "Параметры фильтра даты"
        type: string
        enum:
          - AFTER
          - BEFORE
          - EQUAL

  WarehouseTransactionFilter:
    type: object
    properties:
      warehouse_item:
        $ref: "#/definitions/Filter"
        description: "Параметры фильтра склада"
      nomenclature_item:
        $ref: "#/definitions/Filter"
        description: "Параметры фильтра номенклатуры"
      period:
        $ref: "#/definitions/PeriodFilter"
        description: "Параметры фильтра начала даты"

paths:
  /reports/formats:
    get:
      summary: Получить список форматов для отчетов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /models:
    get:
      summary: Получить список всех моделей
      operationId: main.get_models
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /reports/{model_name}/{format_str}:
    get:
      summary: Получить список единиц измерения
      operationId: main.reports
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    parameters:
      - name: format_str
        in: path
        description: Формат
        required: true
        type: string
      - name: model_name
        in: path
        description: Модель, для которой необходим отчет
        required: true
        type: string

  /{entity}/filter:
    post:
      summary: Фильтрация моделей
      operationId: main.filter_model
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: filter
          description: Параметры фильтрации для номенклатур
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Filter'
        - in: path
          name: entity
          description: Параметры фильтрации для номенклатур
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /api/warehouse_transactions:
    post:
      summary: Транзакции склада
      operationId: main.warehouse_transactions
      produces:
        - application/json
      parameters:
        - in: body
          name: filterDto
          description: Параметры фильтрации для номенклатур
          required: false
          schema:
            type: array
            items:
              $ref: '#/definitions/WarehouseTransactionFilter'
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /api/warehouse_turnovers:
    post:
      summary: Обороты склада
      operationId: main.warehouse_turnovers
      produces:
        - application/json
      parameters:
        - in: body
          name: filterDto
          description: Параметры фильтрации для номенклатур
          required: false
          schema:
            type: array
            items:
              $ref: '#/definitions/WarehouseTransactionFilter'
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /api/date_block:
    get:
      summary: Получить текущую дату блокировки
      operationId: main.get_dateblock
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    post:
      summary: Изменить дату блокировки
      operationId: main.set_dateblock
      produces:
        - application/json
      parameters:
        - in: body
          name: dateblock
          description: Новая дата блокировки
          required: true
          schema:
            type: object
            properties:
              dateblock:
                type: string
                format: date
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка

  /api/nomenclature/{id}:
    get:
      summary: Получить номенклатуру
      operationId: main.get_nomenclature
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID номенклатуры
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    delete:
      summary: Удалить номенклатуру
      operationId: main.delete_nomenclature
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          description: ID номенклатуры
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  /api/nomenclature:
    put:
      summary: Добавить номенклатуру
      operationId: main.put_nomenclature
      produces:
        - application/json
      parameters:
        - in: body
          name: nomenclature
          description: Номенклатура
          required: true
          schema:
            $ref: '#/definitions/NomenclatureModel'
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
    patch:
      summary: Изменить номенклатуру
      operationId: main.patch_nomenclature
      produces:
        - application/json
      parameters:
        - in: body
          name: nomenclature
          description: Номенклатура
          required: true
          schema:
            $ref: '#/definitions/NomenclatureModel'
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  /api/balance_sheet:
    get:
      summary: Получить оборотно-сальдовую ведомость
      operationId: main.get_balance_sheet
      produces:
        - application/json
      parameters:
        - in: body
          name: nomenclature
          description: Номенклатура
          required: true
          schema:
            $ref: '#/definitions/NomenclatureModel'
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  /api/save_data:
    get:
      summary: Сохранить все данные
      operationId: main.save_data
      produces:
        - application/json
      parameters:
        - in: body
          name: filename
          description: Название файла
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  /api/load_data:
    get:
      summary: Загрузить данные
      operationId: main.load_data
      produces:
        - application/json
      parameters:
        - in: body
          name: filename
          description: Название файла
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          schema:
            type: object
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка